refactor(database): split database entrypoints and type surfaces
